+++ libraries/arrays/src/pickItems.ts
@@ -1,4 +1,5 @@
 import { shuffle } from './shuffle.ts';
+import { Seed } from '../../numbers/src/numbers.types.ts';
 
 /**
  * Returns an array with `n` items picked randomly from the given array.
@@ -17,5 +18,5 @@ export function pickItems<T>(items: ReadonlyArray<T>, n: number, options: Option
 
 type Options = {
   offset?: number | undefined;
-  seed?: number | undefined;
+  seed?: Seed;
 };