+++ libraries/numbers/mod.ts
@@ -1,4 +1,7 @@
 export { createNumberGenerator } from './src/createNumberGenerator.ts';
+export { makeSeededRandom } from './src/makeSeededRandom.ts';
 export { pickInteger } from './src/pickInteger.ts';
 export { random } from './src/random.ts';
 export { scale } from './src/scale.ts';
+export { spawnSeedFunction } from './src/spawnSeedFunction.ts';
+export * from './types.ts';