+++ libraries/numbers/src/evaluate.ts
@@ -1,0 +1,5 @@
+export function evaluate<T>(input: FunctionOrValue<T>): T {
+  return input instanceof Function ? input() : input;
+}
+
+export type FunctionOrValue<T> = T | (() => T);