+++ libraries/numbers/src/random.ts
@@ -1,11 +1,14 @@
 import { scale } from './scale.ts';
 import { createNumberGenerator } from './createNumberGenerator.ts';
+import { Seed } from './numbers.types.ts';
 
 /**
- * Returns a random number in the range [min, max).
+ * Returns a scaled random number in the range [min, max).
  */
 export function random(options: Options = {}) {
   const { min = 0, max = 1, seed } = options;
+
+  // Generates a deterministic result if seed is provided; otherwise, uses Math.random().
   const random = createNumberGenerator(seed);
 
   return scale(random(), { min, max });
@@ -14,5 +17,5 @@ export function random(options: Options = {}) {
 interface Options {
   max?: number;
   min?: number;
-  seed?: number;
+  seed?: Seed;
 }