+++ libraries/strings/mod.ts
@@ -1,4 +1,5 @@
 export { capitalize } from './src/capitalize.ts';
+export { concatenate } from './src/concatenate.ts';
 export { Interpolable } from './src/Interpolable.ts';
 export { interpolate } from './src/interpolate.ts';
 export { obfuscate } from './src/obfuscate.ts';
@@ -7,3 +8,8 @@ export { slugify } from './src/slugify.ts';
 export { toCamelCase } from './src/toCamelCase.ts';
 export { unindent } from './src/unindent.ts';
 export { validateDelimiters } from './src/validateDelimiters.ts';
+
+export * from './src/String-prototype-wrappers.ts';
+export * from './src/whitespace-functions.ts';
+
+export * from './types.ts';